Experiment with using preload script for Context Isolation
Electron is moving away from allowing rendere process to have access to node. So I was learning how that works and getting it to work with typescript.

I had not yet actually implemented the context isolation.

However after more thought... lameta is so heavy with node filesystem work that it would be super expensive to wrap every fs access in a call to the main process. And since lameta does not allow JS execution, it not actually a security threat. So I decided it's better to abandoning this work and stick with calling ipc directly from the renderer process.

Still, this is important sample code. I may implement this in another project, just not lameta.